Moved scopes_from_paths from environment.py to config.py
