+++ lib/spack/spack/environment/environment.py
@@ -11,7 +11,7 @@ import re
 import shutil
 import stat
 import warnings
-from typing import Any, Callable, Dict, Iterable, List, Optional, Sequence, Tuple, Union
+from typing import Any, Dict, Iterable, List, Optional, Sequence, Tuple, Union
 
 import llnl.util.filesystem as fs
 import llnl.util.tty as tty
@@ -39,11 +39,9 @@ import spack.util.lock as lk
 import spack.util.path
 import spack.util.spack_json as sjson
 import spack.util.spack_yaml as syaml
-import spack.util.url
 from spack import traverse
 from spack.installer import PackageInstaller
 from spack.schema.env import TOP_LEVEL_KEY, schema, update
-from spack.schema.merged import schema as merged_schema
 from spack.spec import Spec
 from spack.spec_list import SpecList
 from spack.util.path import substitute_path_variables
@@ -2961,7 +2959,7 @@ class EnvironmentManifestFile(collections.abc.Mapping):
             cfg_path = os.path.join(self.manifest_dir, cfg_path)
             return os.path.normpath(os.path.realpath(cfg_path))
 
-        return scopes_from_paths(
+        return spack.config.scopes_from_paths(
             include_paths,
             f"env:{self.name}",
             self.config_stage_dir,
@@ -3028,117 +3026,6 @@ def environment_path_scopes(name: str, path: str) -> Optional[List[spack.config.
         scope.writable = False
 
     return manifest.env_config_scopes
-
-
-# Track staged paths so can avoid repeating the warning during processing
-skip_restage_warning = list()
-
-
-def scopes_from_paths(
-    includes: List[str],
-    name_prefix: str,
-    config_stage_dir: str,
-    resolve_relative: Callable[[str], str],
-    required: List[str] = [],
-) -> List[spack.config.ConfigScope]:
-    """Load included config scopes
-    Scopes are added in reverse order so that highest-precedence scopes are last.
-    Args:
-        includes: list of paths to be included
-        name_prefix: environment's name prefix
-        config_stage_dir: path to directory to be used to stage remote includes
-        resolve_relative: method to use to resolve relative paths
-        required: list of paths that are required
-    Raises:
-        ValueError: included path has an unsupported URL scheme or is required
-            but does not exist
-    """
-    import urllib.parse
-    import urllib.request
-
-    scopes: List[spack.config.ConfigScope] = []
-
-    missing = []
-    for i, orig_path in enumerate(reversed(includes)):
-        # allow paths to contain spack config/environment variables, etc.
-        config_path = substitute_path_variables(orig_path)
-        include_url = urllib.parse.urlparse(config_path)
-
-        # If scheme is not valid, config_path is not a url
-        # of a type Spack is generally aware
-        if spack.util.url.validate_scheme(include_url.scheme):
-            # Transform file:// URLs to direct includes.
-            if include_url.scheme == "file":
-                config_path = urllib.request.url2pathname(include_url.path)
-
-            # Any other URL should be fetched.
-            elif include_url.scheme in ("http", "https", "ftp"):
-                # Stage any remote configuration file(s)
-                staged_configs = (
-                    os.listdir(config_stage_dir) if os.path.exists(config_stage_dir) else []
-                )
-                remote_path = urllib.request.url2pathname(include_url.path)
-                basename = os.path.basename(remote_path)
-                if basename in staged_configs:
-                    # Do NOT re-stage configuration files over existing
-                    # ones with the same name since there is a risk of
-                    # losing changes (e.g., from 'spack config update').
-                    if remote_path not in skip_restage_warning:
-                        tty.warn(
-                            f"Will not re-stage configuration from {remote_path} "
-                            "to avoid losing changes to the already staged file of "
-                            "the same name."
-                        )
-                        skip_restage_warning.append(remote_path)
-
-                    # Recognize the configuration stage directory
-                    # is flattened to ensure a single copy of each
-                    # configuration file.
-                    config_path = config_stage_dir
-                    if basename.endswith(".yaml"):
-                        config_path = os.path.join(config_path, basename)
-                else:
-                    staged_path = spack.config.fetch_remote_configs(
-                        config_path, str(config_stage_dir), skip_existing=True
-                    )
-                    if not staged_path:
-                        raise SpackEnvironmentError(
-                            f"Unable to fetch remote configuration {config_path}"
-                        )
-                    config_path = staged_path
-
-            elif include_url.scheme:
-                raise ValueError(
-                    f"Unsupported URL scheme ({include_url.scheme}) for "
-                    f"environment include: {config_path}"
-                )
-
-        # treat relative paths as relative to the environment
-        if not os.path.isabs(config_path):
-            config_path = resolve_relative(config_path)
-
-        if os.path.isdir(config_path):
-            # directories are treated as regular ConfigScopes
-            config_name = f"{name_prefix}:{os.path.basename(config_path)}"
-            tty.debug(f"Creating DirectoryConfigScope {config_name} for '{config_path}'")
-            scopes.append(spack.config.DirectoryConfigScope(config_name, config_path))
-        elif os.path.exists(config_path):
-            # files are assumed to be SingleFileScopes
-            config_name = f"{name_prefix}:{config_path}"
-            tty.debug(f"Creating SingleFileScope {config_name} for '{config_path}'")
-            scopes.append(spack.config.SingleFileScope(config_name, config_path, merged_schema))
-        elif orig_path in required:
-            raise ValueError(f"Required include path does not exist: {orig_path}")
-        else:
-            missing.append(config_path)
-            continue
-
-    if missing:
-        msg = f"Detected {len(missing)} optional missing include path(s):"
-        msg += "\n   {0}".format("\n   ".join(missing))
-        tty.warn(msg)
-
-    return scopes
 
 
 class SpackEnvironmentError(spack.error.SpackError):
